requirements.txt
fastapi==0.99.0
uvicorn==0.23.0
mysql-connector-python==8.0.33
pydantic>=1.10,<2.0
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]